chore: split mural service internals
